even more minor minor changes

+++ src/main/java/org/sparky/sparkyai/company/domain/CompanyService.java
@@ -64,7 +64,7 @@ public class CompanyService {
         CompanyConsumptionDto report = new CompanyConsumptionDto();
         report.setCallHistory(calls);
         report.setTotalCalls(calls.size());
-        report.setTotalConsumedTokens(calls.stream().map(call -> call.getConsumedTokens()).reduce(0L, Long::sum));
+        report.setTotalConsumedTokens(calls.stream().map(call -> call.getConsumedTokens()).reduce(0, Integer::sum));
         return report;
     }
